nvdimm: selectively alias the data region for PPC
The patch adds new callbacks to nvdimm class structure
for bind/unbind calls and one public callback to memory structure
to set the alias.

For x86 set the alias as soon as the pre_plug checks
and never unalias. For ppc, dont alias the memory region during
plug, do it during the hcall and unalias during machine reset
for the driver bind request to pass.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>